Split ResolutionExecutor into DefaultConfigurationResolver
Previously, ResolutionExecutor implemented ConfigurationResolver. However, the intention is for ResolutionExecutor to be independent of a Configuration.
It should be able to perform any arbitrary resolution, regardless of the DSL construct used to set up the resolution.

We split out all logic that is specific to resolving a Configuration into a DefaultConfigurationResolver, then ResolutionExecutor no longer implements that interface.
DefaultConfigurationResolver now delegates to ResolutionExecutor after it determines the proper repositories to resolve from.

There is an added advantage that constructing a ResolutionExecutor becomes much simpler, as we now just register its class in the service registry instead of calling its
constructor with many parameters.
